Moved ValueIndex into its own ast2ram file.
